requirements.txt
streamlit
python-dotenv
langchain-community 
langchain 
langchain-core 
langchain-openai
crewai 
